capstone tenmo

+++ java/tenmo-server/src/main/java/com/techelevator/tenmo/model/Account.java
@@ -4,14 +4,14 @@ import java.math.BigDecimal;
 
 public class Account {
 
-    private Long accountId;
-    private Long userId;
+    private int accountId;
+    private int userId;
     private BigDecimal balance;
 
-    public Long getAccountId(){
+    public int getAccountId(){
         return accountId;
     }
-    public Long getUserId(){
+    public int getUserId(){
         return userId;
     }
 
@@ -19,7 +19,7 @@ public class Account {
         return balance;
     }
 
-    public void setAccountId(Long accountId){
+    public void setAccountId(int accountId){
         this.accountId = accountId;
     }
 
@@ -27,7 +27,7 @@ public class Account {
         this.balance = balance;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
 }
